* Support large tx by default - remove configuration option * Fix compilation issues * Test LZ4 - cast to int sizes (TODO: Add big size LZ4 compress/uncompress test) * TODO: Discuss how to handle max scratch buffer

+++ bench/Regression.Benchmark/CompressionBench.cs
@@ -43,7 +43,7 @@ namespace Regression
                 fixed (byte* encodedOutputPtr = encodedOutput)
                 fixed (byte* outputPtr = input)
                 {
-                    int compressedSize = lz4.Encode64(inputPtr, encodedOutputPtr, input.Length, maximumOutputLength);
+                    int compressedSize = lz4.Encode64(inputPtr, encodedOutputPtr, input.Length, (int)maximumOutputLength); // ADIADI
                     int uncompressedSize = LZ4.Decode64(encodedOutputPtr, compressedSize, outputPtr, input.Length, true);
                 }
             });
@@ -68,7 +68,7 @@ namespace Regression
                 fixed (byte* encodedOutputPtr = encodedOutput)
                 fixed (byte* outputPtr = input)
                 {
-                    int compressedSize = lz4.Encode64(inputPtr, encodedOutputPtr, input.Length, maximumOutputLength);
+                    int compressedSize = lz4.Encode64(inputPtr, encodedOutputPtr, input.Length, (int)maximumOutputLength); // ADIADI
                     int uncompressedSize = LZ4.Decode64(encodedOutputPtr, compressedSize, outputPtr, input.Length, true);
                 }
             });
